Move Clickable and Link components out of common.js
