+++ src/components/common.js
@@ -7,13 +7,14 @@ import { b, div, h, h1, img, input, label, span } from 'react-hyperscript-helper
 import RSelect, { components as RSelectComponents } from 'react-select'
 import RAsyncCreatableSelect from 'react-select/async-creatable'
 import RSwitch from 'react-switch'
+import Clickable from 'src/components/common/Clickable'
+import Link from 'src/components/common/Link'
 import FooterWrapper from 'src/components/FooterWrapper'
-import { centeredSpinner, containsUnlabelledIcon, icon } from 'src/components/icons'
+import { centeredSpinner, icon } from 'src/components/icons'
 import { TextInput } from 'src/components/input'
 import Interactive from 'src/components/Interactive'
 import Modal from 'src/components/Modal'
 import { MiniSortable } from 'src/components/table'
-import TooltipTrigger from 'src/components/TooltipTrigger'
 import TopBar from 'src/components/TopBar'
 import landingPageHero from 'src/images/landing-page-hero.jpg'
 import scienceBackground from 'src/images/science-background.jpg'
@@ -28,6 +29,8 @@ import { forwardRefWithName, useCancellation, useLabelAssert, useOnMount, useUni
 import { authStore } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
 
+
+export { Clickable, Link }
 
 const styles = {
   button: {
@@ -53,54 +56,6 @@ const styles = {
     }
   }
 }
-
-export const Clickable = forwardRefWithName('Clickable',
-  ({ href, as = (!!href ? 'a' : 'div'), disabled, tooltip, tooltipSide, tooltipDelay, useTooltipAsLabel, onClick, children, ...props }, ref) => {
-    const child = h(Interactive, {
-      'aria-disabled': !!disabled,
-      as, disabled, ref,
-      onClick: (...args) => onClick && !disabled && onClick(...args),
-      href: !disabled ? href : undefined,
-      tabIndex: disabled ? '-1' : '0',
-      ...props
-    }, [children])
-
-    // To support accessibility, every link must have a label or contain text or a labeled child.
-    // If an unlabeled link contains just a single unlabeled icon, then we should use the tooltip as the label,
-    // rather than as the description as we otherwise would.
-    //
-    // If the auto-detection can't make the proper determination, for example, because the icon is wrapped in other elements,
-    // you can explicitly pass in a boolean as `useTooltipAsLabel` to force the correct behavior.
-    //
-    // Note that TooltipTrigger does this same check with its own children, but since we'll be passing it an
-    // Interactive element, we need to do the check here instead.
-    const useAsLabel = _.isNil(useTooltipAsLabel) ? containsUnlabelledIcon({ children, ...props }) : useTooltipAsLabel
-
-    // If we determined that we need to use the tooltip as a label, assert that we have a tooltip.
-    // Do the check here and pass empty properties, to bypass the check logic in useLabelAssert() which doesn't take into account the icon's properties.
-    if (useAsLabel && !tooltip) {
-      useLabelAssert('Clickable', { allowTooltip: true, allowContent: true })
-    }
-
-    if (tooltip) {
-      return h(TooltipTrigger, { content: tooltip, side: tooltipSide, delay: tooltipDelay, useTooltipAsLabel: useAsLabel }, [child])
-    } else {
-      return child
-    }
-  })
-
-export const Link = forwardRefWithName('Link', ({ disabled, variant, children, baseColor = colors.accent, ...props }, ref) => {
-  return h(Clickable, _.merge({
-    ref,
-    style: {
-      color: disabled ? colors.disabled() : baseColor(variant === 'light' ? 0.3 : 1),
-      cursor: disabled ? 'not-allowed' : 'pointer',
-      fontWeight: 500, display: 'inline'
-    },
-    hover: disabled ? undefined : { color: baseColor(variant === 'light' ? 0.1 : 0.8) },
-    disabled
-  }, props), [children])
-})
 
 export const ButtonPrimary = ({ disabled, danger = false, children, ...props }) => {
   return h(Clickable, _.merge({
